moves movieReqHandler() to movies.js and weatherReqHandler() to weather.js
