+++ routeHandlers/movies.js
@@ -9,7 +9,6 @@ async function getMovieData(city_name) {
   const key = 'movie-' + city_name;
   const moviedbUrl = `https://api.themoviedb.org/3/search/movie?api_key=${process.env.MOVIE_API_KEY}&query=${city_name}&adult=false`;
 
-  console.log('cache: ', cache);
   // check if data for query is in cache and up to date by 1 day
   if (cache[key] && (Date.now() - cache[key].timestamp < (3600 * 1000 * 24))) {
     console.log('Cache data found');
@@ -18,9 +17,6 @@ async function getMovieData(city_name) {
     cache[key] = {}; // add obj to key in cache
     cache[key].timestamp = Date.now(); // add current date/time to timestamp in cache
     cache[key].data = parseMovieData(await axios.get(moviedbUrl)); // get current data from movie db
-    // console.log('Axios Movie Data: ', await axios.get(moviedbUrl));
-    // console.log('cache[key].data: ', cache);
-    // console.log('cache[key].data: ', cache[key].data);
   }
 
   return cache[key].data;
@@ -28,8 +24,21 @@ async function getMovieData(city_name) {
 
 function parseMovieData(movieData) {
   const movie = movieData.data.results.map(day => new Movie(day));
-  console.log('parsedMovie: ', movie);
   return movie;
+}
+
+// handle req from client, pass to movies.js, send res to client
+async function movieReqHandler(req, res) {
+  try {
+    // assign lat lon from query to variables
+    const { city_name } = req.query;
+    // pass lat and lon into getWeather() in weather.js to get data from cache or new axios request
+    const movieData = await getMovieData(city_name);
+    res.status(200).send(movieData);
+  } catch (error) {
+    // add error handler
+    res.status(500).send(`There was an error retrieving movie data for ${req.query.city_name}.`);
+  }
 }
 
 class Movie {
@@ -44,4 +53,4 @@ class Movie {
   }
 }
 
-module.exports = getMovieData;
+module.exports = movieReqHandler;
